+++ src/app.module.ts
@@ -4,6 +4,8 @@ import { AppService } from './app.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { CustomerModule } from './customer/customer.module';
 import { Customer } from './customer/entities/customer.entity';
+import { StoreModule } from './store/store.module';
+import { Store } from './store/entities/store.entity';
 
 @Module({
   imports: [TypeOrmModule.forRoot({
@@ -12,9 +14,9 @@ import { Customer } from './customer/entities/customer.entity';
     port: 5432,
     password: '1234',
     username: 'postgres',
-    entities: [Customer],
+    entities: [Customer, Store],
     database: 'commerce'
-    }), CustomerModule],
+    }), CustomerModule, StoreModule],
   controllers: [AppController],
   providers: [AppService],
 })