+++ src/store/dto/update-store.dto.ts
@@ -1,0 +1,5 @@
+import { CreateStoreDto } from './create-store.dto';
+
+export class UpdateStoreDto extends CreateStoreDto {
+    
+}