+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # CHANGELOG
 
-## 02/03/2026
+## 02/10/2026
 - Add customer module
 - Add customer service
 - Add customer controller
 - Add customer DTOs
+
+## 02/11/2026
+- Add store module
+- Add store service
+- Add store controller
+- Add store DTOs